requirements.txt
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pymongo==4.6.1
redis==5.0.1
kafka-python==2.0.2
pyspark==3.5.0
scikit-learn==1.4.0
xgboost==2.0.3
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
streamlit==1.30.0
plotly==5.18.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
faker==22.0.0
pytest==7.4.4
black==24.1.1
flake8==7.0.0